Clean: Relocate STEX LineBoundedByPolygon
