+++ src/dewarping/CMakeLists.txt
@@ -19,7 +19,6 @@ SET(
         STEX_ConnectivityMap.cpp STEX_ConnectivityMap.h
         STEX_GaussBlur.cpp STEX_GaussBlur.h
         STEX_HomographicTransform.h
-        STEX_LineBoundedByPolygon.cpp STEX_LineBoundedByPolygon.h
         STEX_PlugHoles.h
         STEX_PolygonRasterizer.cpp STEX_PolygonRasterizer.h
         STEX_RasterOpGeneric.h
