+++ src/dewarping/DewarpingImageTransform.cpp
@@ -21,7 +21,7 @@
 #include "foundation/RoundingHasher.h"
 #include "STEX_ToVec.h"
 #include "ToLineProjector.h"
-#include "STEX_LineBoundedByPolygon.h"
+#include "math/LineBoundedByPolygon.h"
 #include "STEX_AffineTransformedImage.h"
 #include <Eigen/Core>
 #include <Eigen/LU>
