+++ src/math/LineBoundedByPolygon.h
